[decorators] Add decorators context and access object.
DecoratorContextObject has the following shape:

{
  kind: ‘class’ | ‘method’ | ‘getter’ | ‘setter’ | ‘field’ | ‘accessor’,
  access: <DecoratorAccessObject>, // present if kind !== ‘class’
  private: true | false, // present if kind !== ‘class’
  static: true | false, // present if kind !== ‘class’
  name: <String>,
  addInitializer: <addInitializerFunction>
}

DecoratorAccessObject has the following shape:
{
  get: (obj)=>obj.name
  set: (obj, value)=>{obj.name = value}
  has: (obj)=>{name in obj}
}

The related strings are internalized by adding them to the
NOT_IMPORTANT_INTERNALIZED_STRING_LIST_GENERATOR macro.

The object definitions are added as extentions of the JSObject class
and their maps are cached in the native context.

The addInitializer, get, set, and has functions are implemented as
Torque builtins.

Bug: 42202709
Change-Id: I4c0febabf9f0f1689f73cf4d5224d4d7d765fc18