[sandbox] Validate parameter count in InterpreterEntryTrampoline
When invoking a JSFunction with Ignition bytecode, there are currently
two separate parameter counts: (1) the parameter count in the
JSDispatchTable entry of the function (used by the caller) and (2) the
parameter count in the BytecodeArray (used by the callee). As the
BytecodeArray is not (yet) referenced through the dispatch table, an
attacker can cause the two values to mismatch (for example by swapping
one BytecodeArray with another one), which would lead to stack
corruption. This CL now fixes that by ensuring that the two values are
identical in the InterpreterEntryTrampoline.

Bug: 40931165, 42204201
Change-Id: I2de063b88b8c338c74ce1ddc28795a0fe929b3e5