[sandbox] Introduce CodeStubArguments::MayHavePaddingArguments
CodeStubArguments are usually used for varargs builtins where the
corresponding JSFunction doesn't have a formal parameter count. In that
case, upon returning from the builtin, only the actual number of
arguments (passed via the kJavaScriptCallArgCountRegister) need to be
popped from the stack as there cannot be any padding arguments. However,
there are some builtins that can be installed on JSFunctions with a
parameter count (and with different parameter counts). Examples include
the generic InterpreterEntryTrampoline, but also the InstantiateAsmJs
builtin or the generic JSToWasm and JSToJS wrappers. In that case, there
may be padding arguments that also need to be removed from the stack.
Previously, the builtins needed to handle this manually, by fetching the
parameter count from the function and using it to compute the total
number of arguments on the stack. As the parameter count was obtained
from the SFI::formal_parameter_count field, this was not safe. With this
CL, we now add support for handling these cases with CodeStubArguments
by allowing the user to specify that there may be additional padding
arguments on the stack and changing CodeStubArguments::PopAndReturn to
take those into account. This CL then migrates the InstantiateAsmJs
builtin to use the new approach for handling padding arguments. A
follow-up CL will do the same for torque builtins.

Bug: 40931165, 348084786
Change-Id: I33f189e4e44de3bbaa17111d870da845e08eb3a5