[wasm][jspi] Replace stack switch lock with CAS
The ExecutionAccess lock was found to be responsible for half of the
overhead of switching to the central stack when calling JS/C++ from a
secondary wasm stack.
Replace it with a single CAS per stack limit. If the CAS fails, we leave
the stack limit to its old value such that we do not overwrite the
special interrupt limit.

R=clemensb@chromium.org

Bug: 355400671
Change-Id: I56985d57ed702bde0d67b6647039d0399c4a0f08