Fix fatal error on breakpoint after wrapped function compile failure
This occurs only in debug builds.
I encountered this issue while using Node.js. In Node.js, there is a process where a source is first compiled as CommonJS using compileFunction. If this compilation fails, the source is then re-registered as a module and recompiled.

When setting a breakpoint in code where the first (CommonJS) compilation fails but the second (module) compilation succeeds, a fatal error occurs in V8 with the following message:

# Fatal error in ../../deps/v8/src/parsing/parser.cc, line 910
# Debug check failed: outer_scope->is_eval_scope().

I have modified the debugger to avoid recompiling scripts that previously failed as wrapped functions.

https://github.com/nodejs/node/issues/55532