[sandbox] Introduce "dynamic" parameter counts in code assembler
All Code assembled through the CodeStubAssembler always has a "static"
parameter count as defined by its call descriptor. However, certain
special builtins also have a "dynamic" parameter count: they are defined
as varargs builtins (the static parameter count is zero) but they can be
installed on JSFunctions with different parameter counts. In those
cases, the builtins previously had to manually determine the correct
parameter count to determine the total number of arguments (including
padding) to be able to remove them from the stack in the epilogue. For
that, they relied on the SFI::formal_parameter_count, and so this was
not sandbox-compatible.

With this CL, we now introduce the concept of "dynamic" parameter counts
in the code assembler and allow these builtins to express (currently via
the CodeStubArguments) that they require it. The actual parameter count
is then loaded via the dispatch table and so is guaranteed to be
correct. Then, CodeStubArguments::PopAndReturn takes this into account
to properly remove all arguments.

This CL then migrates the InstantiateAsmJs builtin to use the new
approach for handling padding arguments. A follow-up CL will do the same
for torque builtins.

Bug: 40931165, 348084786
Change-Id: I33f189e4e44de3bbaa17111d870da845e08eb3a5