Use Mutex::AssertHeld instead of TryLock
The pattern `DCHECK(!mutex.TryLock());` is problematic in builds that
continue after failed DCHECKs. We would have locked the mutex one more
time, which can result in deadlocks.
Replace it by the safe `mutex.AssertHeld()`.

This CL also adds a few comments about the race-freeness of
`Mutex::level_` and `RecursiveMutex::level_`.
It also removes `Mutex::AssertUnheld` which was never called and would
result in race conditions if being used.

R=jkummerow@chromium.org

Change-Id: I915e8a2f4e1af8f8c3ffcbf7015e45b4408ddd34